+++ lizard_apps/urls.py
@@ -6,11 +6,11 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from django.conf.urls import url
+from django.urls import re_path
 
 from lizard_apps import views
 
 urlpatterns = [
-    url(r"^screens/(?P<slug>[\w-]+)\.json$", views.app_screen_view),
-    url(r"^screens/(?P<slug>[\w-]+)\.js$", views.AppScreenView.as_view()),
+    re_path(r"^screens/(?P<slug>[\w-]+)\.json$", views.app_screen_view),
+    re_path(r"^screens/(?P<slug>[\w-]+)\.js$", views.AppScreenView.as_view()),
 ]